Add the test to regress_listener.c
