+++ CMakeLists.txt
@@ -1387,7 +1387,6 @@ if (NOT EVENT__DISABLE_TESTS)
                   test-fdleak
                   test-init
                   test-time
-                  test-reuseport-unix
                   test-weof)
 
     foreach (TESTPROG ${TESTPROGS} test-dumpevents)
